# No heap allocations