@ryanslade 2019-05-10
Thanks for this, however I'd rather not introduce the use of the `unsafe` package. 

Also, there is a chance that a user agent contains non ascii characters which this change ignores.