@erikdubbelboer 2019-05-10
Yes the user agent might contains non ascii characters but those aren't used in any of the matching so it wouldn't matter if they are lower cased or not.

What's the reason for not wanting to introduce the use of the `unsafe` package?