@dzbarsky 2022-06-22
@erikdubbelboer It seems like you can use `strings.Builder` instead of an `[]byte` and then you don't need the unsafe string/[]byte conversion. Also, you can create a parser struct that has the strings.Builder on it to reuse the allocation, then you wouldn't need the sync.Pool.