@erikdubbelboer 2022-06-22
This is about the string that is lowercased. `strings.Builder` has nothing to do with it. Besides `strings.Builder` always causes an allocation which is what this pull request is trying to avoid.